MINOTAUR.jl

Documentation for MINOTAUR.jl

MINOTAUR.MINOTAUR — Module

This is the MINOTAUR, it contains

MINOTAUR.MINOTAUR

MINOTAUR.logistic

MINOTAUR.f_logistic

MINOTAUR.g_sqrtquad

MINOTAUR.g_sqrtquad_inv_pos

MINOTAUR.g_sqrtquad_inv_neg

MINOTAUR.interval

MINOTAUR.quadrature_fg

MINOTAUR.τ_root_sphere

MINOTAUR.quadrature_fg_sphere

MINOTAUR.quadrature_fg_sphere_g_opt_f

MINOTAUR.τ_root_cylinder

MINOTAUR.quadrature_fg_cylinder

MINOTAUR.quadrature_fg_cylinder_g_opt_f

Attempt to optimize the discretization for the quadrature to compute the integral

    ∫_τ1^τ2 f(g(τ)) dτ

assuming that g is a bijection, and that g^{-1} assumes non-infinite values

We choose to discretize τ so that enough details of f are represented

We choose a subdivision r_i that captures the details of f over the interval [min(g(τ1),g(τ2)),max(g(τ1),g(τ2))]

MINOTAUR.logistic — Function

logistic(x::Cdouble,thX::Cdouble=33.275)

returns the value of the logistic function at x, 0 if x<-thX, 1 if x>thX
    1/(1+e^{-x})

logistic(x::Cdouble,a::Cdouble,b::Cdouble,α::Cdouble=1.0,thX::Cdouble=33.275)

returns the value of the logistic function at x, 0 if αx<-thX, 1 if αx>thX
    a + (b-a)/(1+(1/α)*e^{-αx})

MINOTAUR.f_logistic — Function

function used for tests of the integral of composition of functions

    f_logistic(r::Array{Cdouble,1},r0::Cdouble,Δr::Cdouble;A::Cdouble=1.0)

input:

  - r: radial distance
  - r0: critical distance 
  - Δr: transition distance
  - A: amplitude (default at 1)

output: 

  - A/(1.0+exp((r-r0)/Δr))

MINOTAUR.g_sqrtquad — Function

g_sqrtquad(τ::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble)

inner function of the composition in the integral ∫_τ1^τ2 f(g(τ)) dτ

Inspired by the distance from a central point to a point on a parametric line 

computes √(a*τ^2+b*τ+c) and should be called with a>0, and b^2<4ac so that a*τ^2+b*τ+c⩾0

MINOTAUR.g_sqrtquad_inv_pos — Function

g_sqrtquad_inv_pos(r::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble)

increasing inverse of sqrt(a*τ^2+b*τ+c), i.e. τ⩾-b/(2a)

should be called with a>0 and b^2<4ac

returns √((r^2 + (b^2/(4a)) - c)/a) - b/(2a)

MINOTAUR.g_sqrtquad_inv_neg — Function

g_sqrtquad_inv_neg(r::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble)

decreasing inverse of sqrt(a*τ^2+b*τ+c), i.e. τ<-b/(2a)

should be called with a>0 and b^2<4ac

returns -√((r^2 + (b^2/(4a)) - c)/a) - b/(2a)

MINOTAUR.interval — Function

interval(τ_min_0::Cdouble,τ_max_0::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble)

Let τ_min=min(τ_min_0,τ_max_0) ⩽ τ_max=max(τ_min_0,τ_max_0)

This function determines the interval boundaries corresponding to g([τ_min,τ_max]) with g(τ) = √(a*τ^2+b*τ+c) (where a>0 and b^2<4ac so that a*τ^2+b*τ+c⩾0)

The determination of the boundaries depends on the values τ_min and τ_max relative to τ0 = -b/(2a) (corrseponding to the argmin of g)

if τ_min and τ_max ⩾ τ0
    g is increasing, so rmin = g(τ_min), and rmax = g(τ_max). rmid = Inf to indicate the lack of intermediate point
if τ_min and τ_max ⩽ τ0
    g is decreasing, so rmin = g(τ_max), and rmax = g(τ_min). rmid = Inf to indicate the lack of intermediate point
if τ_min⩽τ0⩽τ_max
    g is decreasing over [τ_min,τ0], and increasing over [τ0,τ_max], rmid = g(τ0), rmin = g(τ_min), and rmax = g(τ_max)

MINOTAUR.quadrature_fg — Function

quadrature_fg(f::Function,τ_min::Cdouble,τ_max::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble,Δr0::Cdouble;Nr_min::Int64=20)

computes the trapezoid quadrature for the integral

    ∫_{τ_min}^{τ_max} f(g(τ)) dτ

with τ_min<τ_max, and g(τ) = √(a*τ^2+b*τ+c) (where a>0 and b^2<4ac so that a*τ^2+b*τ+c⩾0)

The minumim resolution for the variable r for f is Δr0, and the discretization of r should have at least Nr_min nodes.

quadrature_fg(fgs::Array{Cdouble,1},τs::Array{Cdouble,1})

computes the quadrature ∑_n f(g(τ_n)) ∫ e_n(τ) dτ ≃ ∫_{τ_min}^{τ_max} f(g(τ)) dτ

with fgs[n] = f(g(τs[n])) and e_n piecewise linear basis functions

MINOTAUR.τ_root_sphere — Function

determines the distance on the parametric line starting from (xm,ym,zm) in the direction (x0,y0,z0)

τ_root_sphere(x0::Cdouble,y0::Cdouble,z0::Cdouble,xm::Cdouble,ym::Cdouble,zm::Cdouble,R2::Cdouble,rm2::Cdouble)

input: 

  - xm, ym and zm: origin of the parametric line (τ=0)
  - x0, y0 and z0: coordinates of a point on the parametric line
  - R2: square radius of the intersecting sphere 
  - rm2: square distance between the center of the sphere and (xm,ym,zm), rm^2=xm^2+ym^2+zm^2  (R2-rm2⩾0)

output: 

  - return the distance between the (xm,ym,zm) and the intersection point (sphere of radius R and straight line between (xm,ym,zm) and (x0,y0,z0))

MINOTAUR.quadrature_fg_sphere — Function

quadrature evaluating the integral ∫_{τmin}^{τmax} f(g(τ)) dτ for a sphere of radius μ0

with g(τ)=√(a*τ^2+b*τ+c) with a, b and c defined for each position in the sphere (depend on the location in the sphere and the direction of the parametric line)

f(r) is a mass function depending depending on the distance from the center of the sphere 

the integral boundaries are such that the parametric curve satifies τmin->(r[n],φ[j],θ[k]) and in τmax->(r(τmax),φ(τmax),θ(τmax)) with r(τmax)=μ0+κΔr 
τmin is set to 0 without loss of generality and τmax is determined for each position in the sphere 

quadrature_fg_sphere(r::Array{Cdouble,1},φ::Array{Cdouble,1},θ::Array{Cdouble,1},x0::Cdouble,y0::Cdouble,z0::Cdouble,μ0::Cdouble,Δr::Cdouble;κ::Cdouble=5.0,Nτ::Int64=20)

input:

  - r,φ,θ: spherical coordinates of the discretization of the sphere (polar axis Oz)
  - x0,y0,z0: cartesian coordinate of a point belonging to the parametric straight line of integration
  - μ0: radius of the sphere 
  - Δr: transition distance of the mass function f

optional input:

  - κ:  maximum radius for the integration is r=μ0+κΔr
  - Nτ: number of discretization nodes 

output:

  - FGQ_rφθ: 3D array containing the quadrature for each discretization nodes (r,φ,θ)

MINOTAUR.quadrature_fg_sphere_g_opt_f — Function

quadrature_fg_sphere_g_opt_f(r::Array{Cdouble,1},θ::Array{Cdouble,1},y::Array{Cdouble,1},x0::Cdouble,y0::Cdouble,z0::Cdouble,μ0::Cdouble,Δr::Cdouble;κ::Cdouble=5.0,Nτ::Int64=20)

quadrature evaluating the integral ∫_{τmin}^{τmax} f(g(τ)) dτ for a sphere of radius μ0 

Contrary to quadrature_fg_sphere the quadrature is adjusted to f instead of g

see quadrature_fg_sphere and quadrature_fg(f::Function,τ_min::Cdouble,τ_max::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble,Δr0::Cdouble;Nr_min::Int64=20)

MINOTAUR.τ_root_cylinder — Function

determines the distance on the parametric line starting from (xm,ym,zm) in the direction (x0,y0,z0) intersecting with the cylinder centered on the Oy axis and of radius R

τ_root_cylinder(x0::Cdouble,y0::Cdouble,z0::Cdouble,xm::Cdouble,ym::Cdouble,zm::Cdouble,R2::Cdouble,rm2::Cdouble)

input: 

  - xm, ym and zm: origin of the parametric line (τ=0)
  - x0, y0 and z0: coordinates of a point on the parametric line
  - R2: square radius of the intersecting cylinder 
  - rm2: square distance between the center of the cylinder and (xm,ym,zm), rm^2=xm^2+ym^2  (R2-rm2⩾0)

output: 

  - return the distance between the (xm,ym,zm) and the intersection point (cylinder of radius R along the Oy axis and straight line between (xm,ym,zm) and (x0,y0,z0))

MINOTAUR.quadrature_fg_cylinder — Function

quadrature evaluating the integral ∫_{τmin}^{τmax} f(g(τ)) dτ for a cylinder of radius μ0 (along the axis Oy)

with g(τ)=√(a*τ^2+b*τ+c) with a, b and c defined for each position in the cylinder (depend on the location in the cylinder and the direction of the parametric line)

f(r) is a mass function depending depending on the distance from the center of the cylinder 

the integral boundaries are such that the parametric curve satifies τmin->(r[n],θ[j],y[k]) and in τmax->(r(τmax),θ(τmax),y(τmax)) with r(τmax)=μ0+κΔr 
τmin is set to 0 without loss of generality and τmax is determined for each position in the cylinder 

quadrature_fg_cylinder(r::Array{Cdouble,1},θ::Array{Cdouble,1},y::Array{Cdouble,1},x0::Cdouble,y0::Cdouble,z0::Cdouble,μ0::Cdouble,Δr::Cdouble;κ::Cdouble=5.0,Nτ::Int64=20)

input:

  - r,θ,y: cylindrical coordinates of the discretization of the cylinder (symmetry axis Oy)
  - x0,y0,z0: cartesian coordinate of a point belonging to the parametric straight line of integration
  - μ0: radius of the cylinder
  - Δr: transition distance of the mass function f

optional input:

  - κ:  maximum radius for the integration is r=μ0+κΔr
  - Nτ: number of discretization nodes 

output:

  - FGQ_rθy: 3D array containing the quadrature for each discretization nodes (r,θ,y)

MINOTAUR.quadrature_fg_cylinder_g_opt_f — Function

quadrature_fg_cylinder_g_opt_f(r::Array{Cdouble,1},θ::Array{Cdouble,1},y::Array{Cdouble,1},x0::Cdouble,y0::Cdouble,z0::Cdouble,μ0::Cdouble,Δr::Cdouble;κ::Cdouble=5.0,Nτ::Int64=20)

quadrature evaluating the integral ∫_{τmin}^{τmax} f(g(τ)) dτ for a cylinder of radius μ0 (along the axis Oy)

Contrary to quadrature_fg_cylinder the quadrature is adjusted to f instead of g

see quadrature_fg_cylinder and quadrature_fg(f::Function,τ_min::Cdouble,τ_max::Cdouble,a::Cdouble,b::Cdouble,c::Cdouble,Δr0::Cdouble;Nr_min::Int64=20)